Implemented more generalised architecture design
The EasyDD architecture design was changed to enable generalised inputs of the cuboid geometry, boundary conditions, and boundary condition storage. Most variables are now organized in structures to facilitate further code generalisation. The variables are then extracted from the structures inside functions. Note that this extraction procedure was not implemeted for every single function; it was not implemented for deeply nested functions in particular.