Added a nonlinear HCP mobility law function and updated some comments.
This nonlinear HCP mobility law performs Newton-Raphson (NR) iteration to solve for velocity. The nonlinear root function is based on the power law drag force seen in problem 10.2.1, pg 214, of Computer Simulations of Dislocations. Note that NR converges slowly for this root function. If the NR iteration fails to converge, the nonlinear system is approximated as the typical linear phonon drag system.